LayoutTests/fast/canvas/webgl/bug-32364.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css"/>
<script src="../../js/resources/js-test-pre.js"></script>
<script src="resources/webgl-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>

description('Regression test for <a href="https://bugs.webkit.org/show_bug.cgi?id=32364">https://bugs.webkit.org/show_bug.cgi?id=32364</a> : <code>[Chromium] SporeViewer demo doesn\'t work in Chromium</code>');

var gl = create3DContext();
var floatProgram = loadProgram(gl, "resources/floatUniformShader.vert", "resources/noopUniformShader.frag");
shouldBeUndefined("gl.useProgram(floatProgram)");
var fval4Loc = gl.getUniformLocation(floatProgram, "fval4");
shouldBe("gl.getError()", "gl.NO_ERROR");
shouldBeUndefined("gl.uniform4fv(fval4Loc, new WebGLFloatArray([0.1, 0.2, 0.4, 1.0]));");

var tmpcanvas = document.createElement("canvas");
tmpcanvas.width = 2;
tmpcanvas.height = 2;
var texture = gl.createTexture();
shouldBeUndefined("gl.bindTexture(gl.TEXTURE_2D, texture)");
shouldBeUndefined("gl.texImage2D(gl.TEXTURE_2D, 0, tmpcanvas)");

successfullyParsed = true;

</script>
<script src="../../js/resources/js-test-post.js"></script>

</body>
</html>
