LayoutTests/fast/canvas/webgl/bug-32364-expected.txt

Regression test for https://bugs.webkit.org/show_bug.cgi?id=32364 : [Chromium] SporeViewer demo doesn't work in Chromium

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl.useProgram(floatProgram) is undefined.
PASS gl.getError() is gl.NO_ERROR
PASS gl.uniform4fv(fval4Loc, new WebGLFloatArray([0.1, 0.2, 0.4, 1.0])); is undefined.
PASS gl.bindTexture(gl.TEXTURE_2D, texture) is undefined.
PASS gl.texImage2D(gl.TEXTURE_2D, 0, tmpcanvas) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
